room-agent/requirements.txt
whisperx
pyannote.audio
torch
scikit-learn
numpy
pandas
soundfile
fastapi
uvicorn
av
scipy
python-multipart
websockets
requests
python-dotenv
